# 后端OpenAPI草案_v1.yaml
openapi: 3.0.3
info:
  title: ChuanYi Backend API (Draft V1)
  version: 1.0.0-draft
  description: |
    Draft API contract for ChuanYi app.
    Scope includes DIY flow, order flow, and added goods/cart flow.
servers:
  - url: http://localhost:8080
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/sku/list:
    get:
      tags: [Component]
      summary: Get component SKU list
      parameters:
        - in: query
          name: categoryId
          schema:
            type: integer
            format: int64
        - in: query
          name: keyword
          schema:
            type: string
        - in: query
          name: pageNo
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseComponentSkuPage'
...
  /api/gallery/list:
    get:
      tags: [Content]
      summary: Get design gallery list
      parameters:
        - in: query
          name: pageNo
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGalleryPage'
  /api/gallery/{id}:
    get:
      tags: [Content]
      summary: Get design gallery detail
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGalleryDetail'
...
  /api/goods/categories:
    get:
      tags: [Goods]
      summary: Get goods categories
      parameters:
        - in: query
          name: withCount
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGoodsCategoryList'
  /api/goods/list:
    get:
      tags: [Goods]
      summary: Get goods SPU list
      parameters:
        - in: query
          name: categoryId
          schema:
            type: integer
            format: int64
        - in: query
          name: keyword
          schema:
            type: string
        - in: query
          name: sortType
          schema:
            type: string
            enum: [DEFAULT, PRICE_ASC, PRICE_DESC, NEWEST]
        - in: query
          name: pageNo
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGoodsSpuPage'
  /api/goods/{spuId}:
    get:
      tags: [Goods]
      summary: Get goods detail
      parameters:
        - in: path
          name: spuId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGoodsDetail'
...
  /api/cart/items:
    post:
      tags: [Cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartAddRequest'
      responses:
        '200':
          description: Added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCartSummary'
  /api/cart/items/{cartItemId}:
    put:
      tags: [Cart]
      summary: Update cart item qty
      parameters:
        - in: path
          name: cartItemId
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartUpdateQtyRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCartSummary'
    delete:
      tags: [Cart]
      summary: Delete cart item
      parameters:
        - in: path
          name: cartItemId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCartSummary'
  /api/cart/items/{cartItemId}/select:
    put:
      tags: [Cart]
      summary: Select or unselect one cart item
      parameters:
        - in: path
          name: cartItemId
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartSelectRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCartSummary'
  /api/cart/items/select-all:
    post:
      tags: [Cart]
      summary: Select or unselect all cart items
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartSelectAllRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCartSummary'
  /api/cart/settlement/preview:
    post:
      tags: [Cart]
      summary: Preview settlement and validation result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettlementPreviewRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseSettlementPreviewData'
...
  /api/admin/goods/list:
    get:
      tags: [Admin]
      summary: List goods for admin
      parameters:
        - in: query
          name: categoryId
          schema:
            type: integer
            format: int64
        - in: query
          name: keyword
          schema:
            type: string
        - in: query
          name: salesStatus
          schema:
            type: string
        - in: query
          name: pageNo
          schema:
            type: integer
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
  /api/admin/goods/{spuId}:
    get:
      tags: [Admin]
      summary: Get goods detail for admin edit
      parameters:
        - in: path
          name: spuId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
...
  /api/admin/component/sku/list:
    get:
      tags: [Admin]
      summary: List component sku
      parameters:
        - in: query
          name: categoryId
          schema:
            type: integer
            format: int64
        - in: query
          name: keyword
          schema:
            type: string
        - in: query
          name: salesStatus
          schema:
            type: string
            enum: [ON_SALE, OFF_SALE]
        - in: query
          name: pageNo
          schema:
            type: integer
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PathId:
      in: path
      name: id
      required: true
      schema:
        type: integer
        format: int64
  schemas:
    ApiResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        requestId:
          type: string
          example: 4f6a3d1f8b5545bb
        data:
          nullable: true
...
    ComponentSku:
      type: object
      properties:
        skuId:
          type: integer
          format: int64
        categoryId:
          type: integer
          format: int64
        name:
          type: string
        specText:
          type: string
        beadDiameterMm:
          type: number
          format: double
        price:
          type: number
          format: double
        stockQty:
          type: integer
        salesStatus:
          type: string
          enum: [ON_SALE, OFF_SALE, OUT_OF_STOCK]
        imageUrl:
          type: string
    PageMeta:
      type: object
      properties:
        pageNo:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
          format: int64
    ApiResponseComponentSkuPage:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                list:
                  type: array
                  items:
                    $ref: '#/components/schemas/ComponentSku'
                page:
                  $ref: '#/components/schemas/PageMeta'
...
    GalleryItem:
      type: object
      properties:
        id:
          type: integer
          format: int64
        snapshotId:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        title:
          type: string
        coverImageUrl:
          type: string
        tags:
          type: array
          items:
            type: string
        authorNickname:
          type: string
        displayStatus:
          type: string
        publishedAt:
          type: string
          format: date-time
    ApiResponseGalleryPage:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                list:
                  type: array
                  items:
                    $ref: '#/components/schemas/GalleryItem'
                page:
                  $ref: '#/components/schemas/PageMeta'
    ApiResponseGalleryDetail:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/GalleryItem'
...
    GoodsCategory:
      type: object
      properties:
        id:
          type: integer
          format: int64
        parentId:
          type: integer
          format: int64
          nullable: true
        name:
          type: string
        iconUrl:
          type: string
        sortNo:
          type: integer
        status:
          type: string
          enum: [ENABLED, DISABLED]
        goodsCount:
          type: integer
    ApiResponseGoodsCategoryList:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                categories:
                  type: array
                  items:
                    $ref: '#/components/schemas/GoodsCategory'
    GoodsSpu:
      type: object
      properties:
        spuId:
          type: integer
          format: int64
        categoryId:
          type: integer
          format: int64
        title:
          type: string
        subtitle:
          type: string
        coverImageUrl:
          type: string
        minPrice:
          type: number
          format: double
        maxPrice:
          type: number
          format: double
        salesStatus:
          type: string
          enum: [ON_SALE, OFF_SALE, SOLD_OUT]
        stockStatus:
          type: string
          enum: [HAS_STOCK, LOW_STOCK, OUT_OF_STOCK]
        tags:
          type: array
          items:
            type: string
    GoodsSku:
      type: object
      properties:
        skuId:
          type: integer
          format: int64
        spuId:
          type: integer
          format: int64
        specText:
          type: string
        skuImageUrl:
          type: string
        price:
          type: number
          format: double
        originPrice:
          type: number
          format: double
        stockQty:
          type: integer
        salesStatus:
          type: string
          enum: [ON_SALE, OFF_SALE]
    ApiResponseGoodsSpuPage:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                list:
                  type: array
                  items:
                    $ref: '#/components/schemas/GoodsSpu'
                page:
                  $ref: '#/components/schemas/PageMeta'
    GoodsDetail:
      type: object
      properties:
        spu:
          $ref: '#/components/schemas/GoodsSpu'
        skus:
          type: array
          items:
            $ref: '#/components/schemas/GoodsSku'
        detailHtml:
          type: string
    ApiResponseGoodsDetail:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/GoodsDetail'
...
    CartCheckoutStatus:
      type: string
      enum:
        - READY
        - BLOCKED_EMPTY
        - BLOCKED_NO_SELECTION
        - BLOCKED_INVALID_ITEMS
        - BLOCKED_NO_ADDRESS
        - BLOCKED_STOCK_CHANGED
    CartSummary:
      type: object
      properties:
        itemCount:
          type: integer
        selectedCount:
          type: integer
        invalidCount:
          type: integer
        totalAmount:
          type: number
          format: double
        selectedAmount:
          type: number
          format: double
        discountAmount:
          type: number
          format: double
        payAmount:
          type: number
          format: double
        checkoutStatus:
          $ref: '#/components/schemas/CartCheckoutStatus'
        checkoutBlockReason:
          type: string
          nullable: true
...
    ApiResponseCartSummary:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                summary:
                  $ref: '#/components/schemas/CartSummary'
    CartAddRequest:
      type: object
      required: [itemType, qty]
      properties:
        itemType:
          type: string
          enum: [NORMAL_GOODS, DIY_SNAPSHOT]
        skuId:
          type: integer
          format: int64
          nullable: true
        snapshotNo:
          type: string
          nullable: true
        qty:
          type: integer
          minimum: 1
          maximum: 999
    CartUpdateQtyRequest:
      type: object
      required: [qty]
      properties:
        qty:
          type: integer
          minimum: 1
          maximum: 999
    CartSelectRequest:
      type: object
      required: [selected]
      properties:
        selected:
          type: boolean
    CartSelectAllRequest:
      type: object
      required: [selected]
      properties:
        selected:
          type: boolean
    SettlementPriceLine:
      type: object
      properties:
        title:
          type: string
        amount:
          type: number
          format: double
    SettlementPreviewRequest:
      type: object
      properties:
        addressId:
          type: integer
          format: int64
        cartItemIds:
          type: array
          items:
            type: integer
            format: int64
    SettlementPreviewData:
      type: object
      properties:
        summary:
          $ref: '#/components/schemas/CartSummary'
        priceLines:
          type: array
          items:
            $ref: '#/components/schemas/SettlementPriceLine'
        canSubmit:
          type: boolean
        blockingReasons:
          type: array
          items:
            type: string
    ApiResponseSettlementPreviewData:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/SettlementPreviewData'